added review requested changes

+++ hw02/include/matrix_calc.cpp
@@ -21,42 +21,42 @@ public:
     }
 
     Matrix(const Matrix<1, M, T>& row_vector) {
-        for (uint row_ind = 0; row_ind < N; ++row_ind)
-            for (uint col_ind = 0; col_ind < M; ++col_ind)
+        for (size_t row_ind = 0; row_ind < N; ++row_ind)
+            for (size_t col_ind = 0; col_ind < M; ++col_ind)
                 array_[row_ind][col_ind] = row_vector[0][col_ind];
     }
 
     Matrix(const Matrix<N, 1, T>& col_vector) {
-        for (uint row_ind = 0; row_ind < N; ++row_ind)
-            for (uint col_ind = 0; col_ind < M; ++col_ind)
+        for (size_t row_ind = 0; row_ind < N; ++row_ind)
+            for (size_t col_ind = 0; col_ind < M; ++col_ind)
                 array_[row_ind][col_ind] = col_vector[row_ind][0];
     }
 
 
     Matrix<N, M, T>& operator+=(const Matrix<N, M, T>& other) {
-        for (uint row_ind = 0; row_ind < N; ++row_ind)
-            for (uint col_ind = 0; col_ind < M; ++col_ind)
+        for (size_t row_ind = 0; row_ind < N; ++row_ind)
+            for (size_t col_ind = 0; col_ind < M; ++col_ind)
                 array_[row_ind][col_ind] += other.array_[row_ind][col_ind];
         return *this;
     }
 
     Matrix<N, M, T>& operator-=(const Matrix<N, M, T>& other) {
-        for (uint row_ind = 0; row_ind < N; ++row_ind)
-            for (uint col_ind = 0; col_ind < M; ++col_ind)
+        for (size_t row_ind = 0; row_ind < N; ++row_ind)
+            for (size_t col_ind = 0; col_ind < M; ++col_ind)
                 array_[row_ind][col_ind] -= other.array_[row_ind][col_ind];
         return *this;
     }
 
     Matrix<N, M, T>& operator*=(const Matrix<N, M, T>& other) {
-        for (uint row_ind = 0; row_ind < N; ++row_ind)
-            for (uint col_ind = 0; col_ind < M; ++col_ind)
+        for (size_t row_ind = 0; row_ind < N; ++row_ind)
+            for (size_t col_ind = 0; col_ind < M; ++col_ind)
                 array_[row_ind][col_ind] *= other.array_[row_ind][col_ind];
         return *this;
     }
 
     Matrix<N, M, T>& operator*=(T scalar) {
-        for (uint row_ind = 0; row_ind < N; ++row_ind)
-            for (uint col_ind = 0; col_ind < M; ++col_ind)
+        for (size_t row_ind = 0; row_ind < N; ++row_ind)
+            for (size_t col_ind = 0; col_ind < M; ++col_ind)
                 array_[row_ind][col_ind] *= scalar;
         return *this;
     }
@@ -76,8 +76,8 @@ public:
             COL_IND_LAST - COL_IND_FIRST,
             T> Slice() const {
         T new_array[ROW_IND_LAST - ROW_IND_FIRST][COL_IND_LAST - COL_IND_FIRST];
-        for (uint row_ind = ROW_IND_FIRST; row_ind < ROW_IND_LAST; ++row_ind)
-            for (uint col_ind = COL_IND_FIRST; col_ind < COL_IND_LAST; ++col_ind)
+        for (size_t row_ind = ROW_IND_FIRST; row_ind < ROW_IND_LAST; ++row_ind)
+            for (size_t col_ind = COL_IND_FIRST; col_ind < COL_IND_LAST; ++col_ind)
                 new_array[row_ind - ROW_IND_FIRST][col_ind - COL_IND_FIRST] =\
                 array_[row_ind][col_ind];
         return new_array;
@@ -86,7 +86,7 @@ public:
     Matrix<1, std::min(N, M), T> Diagonal() const {
         size_t n_m_min = std::min(N, M);
         T row_vector[1][std::min(N, M)];
-        for (uint ind = 0; ind < n_m_min; ++ind)
+        for (size_t ind = 0; ind < n_m_min; ++ind)
             row_vector[0][ind] = array_[ind][ind];
         return row_vector;
     }
@@ -94,8 +94,8 @@ public:
 
     Matrix<M, N, T> Transpose() const {
         T new_array[M][N];
-        for (uint row_ind = 0; row_ind < N; ++row_ind)
-            for (uint col_ind = 0; col_ind <M; ++col_ind)
+        for (size_t row_ind = 0; row_ind < N; ++row_ind)
+            for (size_t col_ind = 0; col_ind <M; ++col_ind)
                 new_array[col_ind][row_ind] = array_[row_ind][col_ind];
         return new_array;
     }
@@ -157,9 +157,9 @@ Matrix<N, M, T> operator*(T first, const Matrix<N, M, T> second) {
 template <size_t N, size_t M, size_t X, typename T>
 Matrix<X, M, T> operator,(const Matrix<X, N, T>& first, const Matrix<N, M, T>& second) {
     Matrix<X, M, T> result(T(0.));
-    for (uint row_ind_first = 0; row_ind_first < X; ++row_ind_first)
-        for (uint col_ind_second = 0; col_ind_second < M; ++col_ind_second)
-            for (uint ind = 0; ind < N; ++ind)
+    for (size_t row_ind_first = 0; row_ind_first < X; ++row_ind_first)
+        for (size_t col_ind_second = 0; col_ind_second < M; ++col_ind_second)
+            for (size_t ind = 0; ind < N; ++ind)
                 result[row_ind_first][col_ind_second] +=\
                 first[row_ind_first][ind] * second[ind][col_ind_second];
     return result;
@@ -169,8 +169,8 @@ template <size_t N, size_t M, typename T, size_t X, size_t Y, typename U>
 bool operator==(const Matrix<N, M, T>& first, const Matrix<X, Y, U>& second) {
     if (N != X || M != Y || !std::is_same_v<T, U>)
         return false;
-    for (uint row_ind = 0; row_ind < N; ++row_ind) {
-        for (uint col_ind = 0; col_ind < M; ++col_ind) {
+    for (size_t row_ind = 0; row_ind < N; ++row_ind) {
+        for (size_t col_ind = 0; col_ind < M; ++col_ind) {
             if (first[row_ind][col_ind] != second[row_ind][col_ind])
                 return false;
         }
@@ -186,7 +186,7 @@ bool operator!=(const Matrix<N, M, T>& first, const Matrix<X, Y, U>& second) {
 template <size_t N, size_t M, typename T>
 std::ostream& operator<<(std::ostream& out, const Matrix<N, M, T>& mat) {
     std::cout << '{' << '\n';
-    for (uint row_col = 0; row_col < N; ++row_col) {
+    for (size_t row_col = 0; row_col < N; ++row_col) {
         std::cout << '{';
         std::copy(mat.array_[row_col], mat.array_[row_col] + M, std::ostream_iterator<T>(out, ", "));
         std::cout << '}' << '\n';
